Unit Testing
Started unit tests, also changed int over to short where possible.

+++ src/main/java/com/wordpress/_0x10f8/logcheck/cli/CommandLineArguments.java
@@ -43,7 +43,7 @@ public class CommandLineArguments {
 	private String outputFile;
 
 	@CommandLine.Option(names = { "-t", "--threads" }, description = "Number of threads to use", required = true)
-	private int threads;
+	private short threads;
 
 	private boolean argumentsSet = false;
 
@@ -145,7 +145,7 @@ public class CommandLineArguments {
 	 * 
 	 * @return number of threads
 	 */
-	public int getThreads() {
+	public short getThreads() {
 		return this.threads;
 	}
